Get re-formations after death working
One consideration is: Do we want to let you try a different pick
as well? Seems a little bit weird to do so.

The play might be, to make getting the imp rewards a totally different
scene. That way, it kind of makes sense that you can't change what
you got.